project/api/code/requirements.txt
cffi==1.14.5
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
# greenlet==0.4.13
gunicorn==20.1.0
importlib-metadata==4.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==2.1.0
readline==6.2.4.1
SQLAlchemy==1.4.15
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
opencv-python-headless==4.5.2.54
face_recognition==1.3.0